refactor(tui): extract render closures into components/
Move TabBar, Footer, ProvidersContent, ProviderDetail, ProfilesContent,
ProfileDetail, RoutingContent, RoutingDetail, PrivacyContent, PrivacyDetail
out of App.tsx into packages/cli/src/tui/components/.

App.tsx state and useKeyboard handler unchanged. Components are pure
props-in-JSX-out. Footer recomputes isProfileEditMode from the `mode` prop
to keep its render contract self-contained. Shared types live in tui/types.ts
and shared constants (COMMON_MODELS, PROVIDER_PREFIXES, CHAIN_PROVIDERS,
layout dims, VERSION) in tui/constants.ts.

Replaces the 3-line stubs in tui/panels/ (deleted).

App.tsx 2,348 -> 959 LOC.
