+++ packages/cli/src/tui/App.tsx
@@ -11,75 +11,32 @@ import {
   setApiKey,
   setEndpoint,
 } from "../profile-config.js";
-import { DEFAULT_ROUTING_RULES } from "../providers/default-routing-rules.js";
 import { route } from "../providers/routing-rules.js";
 import { clearBuffer, getBufferStats } from "../stats-buffer.js";
 import { testProviderKey } from "./test-provider.js";
-import { PROVIDERS, ProviderDef, maskKey } from "./providers.js";
+import { PROVIDERS, maskKey } from "./providers.js";
 import { C } from "./theme.js";
-
-const VERSION = "v5.16";
-
-// ── Common models for autocomplete ────────────────────────────────────────────
-const COMMON_MODELS = [
-  "g@gemini-3.1-pro-preview",
-  "g@gemini-2.5-flash",
-  "g@gemini-2.5-pro",
-  "oai@gpt-4o",
-  "oai@gpt-4o-mini",
-  "oai@o3-mini",
-  "or@anthropic/claude-sonnet-4-20250514",
-  "mm@minimax-m2.5",
-  "kimi@kimi-k2.5",
-  "glm@glm-5",
-  "zen@glm-5",
-  "zen@minimax-m2.5-free",
-  "ll@gemini-2.5-flash",
-  "ll@gpt-4o",
-  "or@google/gemini-3.1-pro-preview",
-  "or@x-ai/grok-code-fast-1",
-  "or@deepseek/deepseek-r1",
-];
-
-// Provider prefix suggestions for the provider picker
-const PROVIDER_PREFIXES = PROVIDERS.map((p) => ({
-  prefix: p.aliases?.[0] ? `${p.aliases[0]}@` : `${p.name}@`,
-  displayName: p.displayName,
-  name: p.name,
-}));
-
-type Tab = "providers" | "profiles" | "routing" | "privacy";
-type Mode =
-  | "browse"
-  | "input_key"
-  | "input_endpoint"
-  | "add_routing_pattern"
-  | "add_routing_chain"
-  | "new_profile"
-  | "pick_profile_scope"
-  | "pick_provider_prefix"
-  | "edit_profile_opus"
-  | "edit_profile_sonnet"
-  | "edit_profile_haiku"
-  | "edit_profile_subagent";
-
-type ProbeMode = "idle" | "input" | "running" | "done";
-
-interface ProbeEntry {
-  provider: string;
-  displayName: string;
-  status: "pending" | "testing" | "success" | "failed" | "skipped" | "no_key";
-  error?: string;
-  ms?: number;
-  hasKey?: boolean;
-  reason?: string;
-}
-
-function bytesHuman(b: number): string {
-  if (b < 1024) return `${b} B`;
-  if (b < 1024 * 1024) return `${(b / 1024).toFixed(1)} KB`;
-  return `${(b / (1024 * 1024)).toFixed(1)} MB`;
-}
+import {
+  COMMON_MODELS,
+  PROVIDER_PREFIXES,
+  CHAIN_PROVIDERS,
+  HEADER_H,
+  TABS_H,
+  FOOTER_H,
+  DETAIL_H,
+  VERSION,
+} from "./constants.js";
+import type { Mode, ProbeEntry, ProbeMode, Tab, TestResultsMap } from "./types.js";
+import { TabBar } from "./components/TabBar.js";
+import { Footer } from "./components/Footer.js";
+import { ProvidersContent } from "./components/ProvidersContent.js";
+import { ProviderDetail } from "./components/ProviderDetail.js";
+import { ProfilesContent } from "./components/ProfilesContent.js";
+import { ProfileDetail } from "./components/ProfileDetail.js";
+import { RoutingContent } from "./components/RoutingContent.js";
+import { RoutingDetail } from "./components/RoutingDetail.js";
+import { PrivacyContent } from "./components/PrivacyContent.js";
+import { PrivacyDetail } from "./components/PrivacyDetail.js";
 
 export function App() {
   const renderer = useRenderer();
@@ -97,9 +54,7 @@ export function App() {
   const [chainOrder, setChainOrder] = useState<string[]>([]);
   const [chainCursor, setChainCursor] = useState(0);
   const [statusMsg, setStatusMsg] = useState<string | null>(null);
-  const [testResults, setTestResults] = useState<
-    Record<string, { status: "testing" | "valid" | "failed"; error?: string; ms?: number }>
-  >({});
+  const [testResults, setTestResults] = useState<TestResultsMap>({});
   const [probeMode, setProbeMode] = useState<ProbeMode>("idle");
   const [probeModel, setProbeModel] = useState("");
   const [probeResults, setProbeResults] = useState<ProbeEntry[]>([]);
@@ -114,9 +69,6 @@ export function App() {
   const [providerPickerIndex, setProviderPickerIndex] = useState(0);
   const [providerPickerReturnMode, setProviderPickerReturnMode] =
     useState<Mode>("edit_profile_opus");
-
-  // Chain selector uses same PROVIDERS list for consistent naming
-  const CHAIN_PROVIDERS = PROVIDERS;
 
   // Compute autocomplete suggestions for model input
   const computeSuggestions = useCallback((input: string): string[] => {
@@ -877,1412 +829,10 @@ export function App() {
 
   const isInputMode = mode === "input_key" || mode === "input_endpoint";
   const isRoutingInput = mode === "add_routing_pattern" || mode === "add_routing_chain";
-  const isProfileEditMode =
-    mode === "new_profile" ||
-    mode === "pick_profile_scope" ||
-    mode === "pick_provider_prefix" ||
-    mode === "edit_profile_opus" ||
-    mode === "edit_profile_sonnet" ||
-    mode === "edit_profile_haiku" ||
-    mode === "edit_profile_subagent";
 
   // ── Layout math ───────────────────────────────────────────────────────────
   // header(1) + tab-bar(3) + content(flex) + detail(fixed) + footer(1)
-  const HEADER_H = 1;
-  const TABS_H = 3;
-  const FOOTER_H = 1;
-  const DETAIL_H = 7;
   const contentH = Math.max(4, height - HEADER_H - TABS_H - DETAIL_H - FOOTER_H - 1);
-
-  // ── Render helpers ────────────────────────────────────────────────────────
-  function TabBar() {
-    const tabs: Array<{ label: string; value: Tab; num: string }> = [
-      { label: "Providers", value: "providers", num: "1" },
-      { label: "Profiles", value: "profiles", num: "2" },
-      { label: "Routing", value: "routing", num: "3" },
-      { label: "Privacy", value: "privacy", num: "4" },
-    ];
-
-    return (
-      <box height={TABS_H} flexDirection="column" backgroundColor={C.bg}>
-        {/* Tab buttons row — use box-level backgroundColor for unmistakable tab highlighting */}
-        <box height={1} flexDirection="row">
-          <box width={1} height={1} backgroundColor={C.bg} />
-          {tabs.map((t, i) => {
-            const active = activeTab === t.value;
-            return (
-              <box key={t.value} flexDirection="row" height={1}>
-                {i > 0 && <box width={2} height={1} backgroundColor={C.bg} />}
-                <box
-                  height={1}
-                  backgroundColor={active ? C.tabActiveBg : C.tabInactiveBg}
-                  paddingX={1}
-                >
-                  <text>
-                    <span fg={active ? C.tabActiveFg : C.tabInactiveFg} bold>
-                      {`${t.num}. ${t.label}`}
-                    </span>
-                  </text>
-                </box>
-              </box>
-            );
-          })}
-          {statusMsg && (
-            <box height={1} backgroundColor={C.bg} paddingX={1}>
-              <text>
-                <span fg={C.dim}>{"─  "}</span>
-                <span
-                  fg={
-                    statusMsg.startsWith("Key saved") ||
-                    statusMsg.startsWith("Rule added") ||
-                    statusMsg.startsWith("Endpoint") ||
-                    statusMsg.startsWith("Telemetry") ||
-                    statusMsg.startsWith("Usage") ||
-                    statusMsg.startsWith("Stats buffer") ||
-                    statusMsg.startsWith("Profile") ||
-                    statusMsg.startsWith("Key removed")
-                      ? C.green
-                      : C.yellow
-                  }
-                  bold
-                >
-                  {statusMsg}
-                </span>
-              </text>
-            </box>
-          )}
-        </box>
-        {/* Separator line */}
-        <box height={1} paddingX={1}>
-          <text>
-            <span fg={C.tabActiveBg}>{"─".repeat(Math.max(0, width - 2))}</span>
-          </text>
-        </box>
-        {/* Spacer */}
-        <box height={1} />
-      </box>
-    );
-  }
-
-  // ── Providers tab ─────────────────────────────────────────────────────────
-  function ProvidersContent() {
-    const listH = contentH - 2; // inner height of box
-    let separatorRendered = false;
-
-    const getRow = (p: ProviderDef, idx: number) => {
-      const isReady = !!(config.apiKeys?.[p.apiKeyEnvVar] || process.env[p.apiKeyEnvVar]);
-      const selected = idx === providerIndex;
-      const cfgMask = maskKey(config.apiKeys?.[p.apiKeyEnvVar]);
-      const envMask = maskKey(process.env[p.apiKeyEnvVar]);
-      const hasCfg = cfgMask !== "────────";
-      const hasEnv = envMask !== "────────";
-      const keyDisplay = isReady ? (hasCfg ? cfgMask : envMask) : "────────";
-      const src = hasEnv && hasCfg ? "e+c" : hasEnv ? "env" : hasCfg ? "cfg" : "";
-      const namePad = p.displayName.padEnd(14).substring(0, 14);
-      const isFirstUnready = !isReady && !separatorRendered;
-      if (isFirstUnready) separatorRendered = true;
-
-      // Inline test result for this provider
-      const tr = testResults[p.name];
-      let statusFg = isReady ? C.green : C.dim;
-      let statusText = isReady ? "ready  " : "not set";
-      if (tr) {
-        if (tr.status === "testing") {
-          statusFg = C.yellow;
-          statusText = "testing";
-        } else if (tr.status === "valid") {
-          statusFg = C.green;
-          statusText = tr.ms !== undefined ? `ready ${tr.ms}ms` : "ready ✓";
-        } else {
-          statusFg = C.red;
-          statusText = "FAIL   ";
-        }
-      }
-
-      return (
-        <box key={p.name} flexDirection="column">
-          {isFirstUnready && (
-            <box height={1} paddingX={1}>
-              <text>
-                <span fg={C.dim}>
-                  {"─ not configured "}
-                  {"─".repeat(Math.max(0, width - 22))}
-                </span>
-              </text>
-            </box>
-          )}
-          <box height={1} flexDirection="row" backgroundColor={selected ? C.bgHighlight : C.bg}>
-            <text>
-              <span fg={tr?.status === "testing" ? C.yellow : isReady ? C.green : C.dim}>
-                {tr?.status === "testing" ? "◌" : isReady ? "●" : "○"}
-              </span>
-              <span>{"  "}</span>
-              <span fg={selected ? C.white : isReady ? C.fgMuted : C.dim} bold={selected}>
-                {namePad}
-              </span>
-              <span fg={C.dim}>{"  "}</span>
-              <span fg={statusFg} bold={tr?.status === "valid" || isReady}>
-                {statusText}
-              </span>
-              <span fg={C.dim}>{"  "}</span>
-              <span fg={isReady ? C.cyan : C.dim}>{keyDisplay}</span>
-              {src ? <span fg={C.dim}>{` (${src})`}</span> : null}
-              <span fg={C.dim}>{"  "}</span>
-              <span fg={selected ? C.white : C.dim}>{p.description}</span>
-            </text>
-          </box>
-        </box>
-      );
-    };
-
-    return (
-      <box
-        height={contentH}
-        border
-        borderStyle="single"
-        borderColor={!isInputMode ? C.blue : C.dim}
-        backgroundColor={C.bg}
-        flexDirection="column"
-        paddingX={1}
-      >
-        {/* Column header */}
-        <text>
-          <span fg={C.dim}>{"   "}</span>
-          <span fg={C.blue} bold>
-            {"PROVIDER        "}
-          </span>
-          <span fg={C.blue} bold>
-            {"STATUS    "}
-          </span>
-          <span fg={C.blue} bold>
-            {"KEY         "}
-          </span>
-          <span fg={C.blue} bold>
-            DESCRIPTION
-          </span>
-        </text>
-        {displayProviders.slice(0, listH).map(getRow)}
-      </box>
-    );
-  }
-
-  function ProviderDetail() {
-    const displayKey = hasCfgKey ? cfgKeyMask : hasEnvKey ? envKeyMask : "────────";
-
-    if (isInputMode) {
-      return (
-        <box
-          height={DETAIL_H}
-          border
-          borderStyle="single"
-          borderColor={C.focusBorder}
-          title={` Set ${mode === "input_key" ? "API Key" : "Endpoint"} — ${selectedProvider.displayName} `}
-          backgroundColor={C.bg}
-          flexDirection="column"
-          paddingX={1}
-        >
-          <text>
-            <span fg={C.green} bold>
-              Enter{" "}
-            </span>
-            <span fg={C.fgMuted}>to save · </span>
-            <span fg={C.red} bold>
-              Esc{" "}
-            </span>
-            <span fg={C.fgMuted}>to cancel</span>
-          </text>
-          <box flexDirection="row">
-            <text>
-              <span fg={C.green} bold>
-                &gt;{" "}
-              </span>
-            </text>
-            <input
-              value={inputValue}
-              onChange={setInputValue}
-              focused={true}
-              width={width - 8}
-              backgroundColor={C.bgHighlight}
-              textColor={C.white}
-            />
-          </box>
-        </box>
-      );
-    }
-
-    const tr = testResults[selectedProvider.name];
-
-    return (
-      <box
-        height={DETAIL_H}
-        border
-        borderStyle="single"
-        borderColor={C.dim}
-        title={` ${selectedProvider.displayName} `}
-        backgroundColor={C.bgAlt}
-        flexDirection="column"
-        paddingX={1}
-      >
-        <box flexDirection="row">
-          <text>
-            <span fg={C.blue} bold>
-              Status:{" "}
-            </span>
-            {hasKey ? (
-              <span fg={C.green} bold>
-                ● Ready
-              </span>
-            ) : (
-              <span fg={C.fgMuted}>○ Not configured</span>
-            )}
-            <span fg={C.dim}>{"    "}</span>
-            <span fg={C.blue} bold>
-              Key:{" "}
-            </span>
-            <span fg={C.green}>{displayKey}</span>
-            {keySrc && <span fg={C.fgMuted}> (source: {keySrc})</span>}
-          </text>
-        </box>
-        {selectedProvider.endpointEnvVar && (
-          <text>
-            <span fg={C.blue} bold>
-              URL:{" "}
-            </span>
-            <span fg={C.cyan}>
-              {activeEndpoint || selectedProvider.defaultEndpoint || "default"}
-            </span>
-          </text>
-        )}
-        <text>
-          <span fg={C.blue} bold>
-            Desc:{" "}
-          </span>
-          <span fg={C.white}>{selectedProvider.description}</span>
-        </text>
-        {selectedProvider.keyUrl && (
-          <text>
-            <span fg={C.blue} bold>
-              Get Key:{" "}
-            </span>
-            <span fg={C.cyan}>{selectedProvider.keyUrl}</span>
-          </text>
-        )}
-        {tr && (
-          <text>
-            <span fg={C.blue} bold>
-              {"Test:  "}
-            </span>
-            {tr.status === "testing" && (
-              <span fg={C.yellow} bold>
-                {"◌ testing..."}
-              </span>
-            )}
-            {tr.status === "valid" && (
-              <>
-                <span fg={C.green} bold>
-                  {"● valid"}
-                </span>
-                {tr.ms !== undefined && <span fg={C.dim}>{`  ${tr.ms}ms`}</span>}
-                <span fg={C.fgMuted}>{"  API key is valid and endpoint is reachable."}</span>
-              </>
-            )}
-            {tr.status === "failed" && (
-              <>
-                <span fg={C.red} bold>
-                  {"✗ failed"}
-                </span>
-                {tr.error && <span fg={C.red}>{`  ${tr.error}`}</span>}
-              </>
-            )}
-          </text>
-        )}
-      </box>
-    );
-  }
-
-  // ── Profiles tab ──────────────────────────────────────────────────────────
-
-  function ProfilesContent() {
-    const globalCfg = config;
-    const localCfg = loadLocalConfig();
-    const localProfileNames = localCfg
-      ? new Set(Object.keys(localCfg.profiles))
-      : new Set<string>();
-
-    // Build unified list: local profiles first, then global
-    const allEntries: Array<{
-      name: string;
-      scope: "local" | "global";
-      models: Record<string, string | undefined>;
-    }> = [];
-    if (localCfg) {
-      for (const [name, prof] of Object.entries(localCfg.profiles)) {
-        allEntries.push({ name, scope: "local", models: prof.models });
-      }
-    }
-    for (const [name, prof] of Object.entries(globalCfg.profiles)) {
-      allEntries.push({ name, scope: "global", models: prof.models });
-    }
-
-    const activeProfileName = globalCfg.defaultProfile;
-    const listH = contentH - 2;
-
-    // Edit mode prompt
-    const editPromptLabel =
-      mode === "new_profile"
-        ? `New ${profileScope} profile — name:`
-        : mode === "pick_profile_scope"
-          ? "Scope for new profile:"
-          : mode === "pick_provider_prefix"
-            ? "Select provider:"
-            : mode === "edit_profile_opus"
-              ? `${editProfileName} — opus model:`
-              : mode === "edit_profile_sonnet"
-                ? `${editProfileName} — sonnet model:`
-                : mode === "edit_profile_haiku"
-                  ? `${editProfileName} — haiku model:`
-                  : mode === "edit_profile_subagent"
-                    ? `${editProfileName} — subagent model (optional):`
-                    : null;
-
-    return (
-      <box
-        height={contentH}
-        border
-        borderStyle="single"
-        borderColor={activeTab === "profiles" && !isProfileEditMode ? C.blue : C.dim}
-        backgroundColor={C.bg}
-        flexDirection="column"
-        paddingX={1}
-      >
-        {/* Active profile indicator */}
-        <text>
-          <span fg={C.dim}>{"  "}</span>
-          <span fg={C.fgMuted}>Active profile: </span>
-          <span fg={C.orange} bold>
-            {activeProfileName}
-          </span>
-        </text>
-        {/* Column header */}
-        <text>
-          <span fg={C.dim}>{"   "}</span>
-          <span fg={C.blue} bold>
-            {"PROFILE         "}
-          </span>
-          <span fg={C.blue} bold>
-            {"SCOPE    "}
-          </span>
-          <span fg={C.blue} bold>
-            {"MODELS"}
-          </span>
-        </text>
-        {/* Profile rows */}
-        {allEntries.slice(0, Math.max(0, listH - 3)).map((entry, idx) => {
-          const isActive = entry.name === activeProfileName;
-          const selected = idx === profileIndex;
-          const namePad = entry.name.padEnd(16).substring(0, 16);
-          const scopePad = entry.scope.padEnd(8).substring(0, 8);
-          const shadowed = entry.scope === "global" && localProfileNames.has(entry.name);
-
-          const modelSummary =
-            [
-              entry.models.opus ? `opus→${entry.models.opus.substring(0, 14)}` : null,
-              entry.models.sonnet ? `sonnet→${entry.models.sonnet.substring(0, 14)}` : null,
-            ]
-              .filter(Boolean)
-              .join("  ") || "(auto-route)";
-
-          return (
-            <box
-              key={`${entry.scope}-${entry.name}`}
-              height={1}
-              flexDirection="row"
-              backgroundColor={selected ? C.bgHighlight : C.bg}
-            >
-              <text>
-                <span fg={isActive ? C.orange : C.dim}>{isActive ? "●" : " "}</span>
-                <span fg={C.dim}> </span>
-                <span
-                  fg={selected ? C.white : isActive ? C.orange : C.fgMuted}
-                  bold={selected || isActive}
-                >
-                  {namePad}
-                </span>
-                <span fg={C.dim}>{"  "}</span>
-                <span fg={entry.scope === "local" ? C.cyan : C.fgMuted}>{scopePad}</span>
-                <span fg={C.dim}>{"  "}</span>
-                <span fg={selected ? C.white : shadowed ? C.dim : C.fgMuted}>
-                  {shadowed ? "(shadowed by local)  " : modelSummary}
-                </span>
-              </text>
-            </box>
-          );
-        })}
-
-        {/* Local profiles note */}
-        {!localCfg && (
-          <text>
-            <span fg={C.dim}>{"  No project-level profiles (.claudish.json)"}</span>
-          </text>
-        )}
-
-        {/* Edit mode input */}
-        {isProfileEditMode && editPromptLabel && (
-          <box flexDirection="column" paddingTop={1}>
-            <text>
-              <span fg={C.blue} bold>
-                {editPromptLabel + " "}
-              </span>
-            </text>
-
-            {/* Scope picker */}
-            {mode === "pick_profile_scope" && (
-              <box flexDirection="column">
-                <box height={1} flexDirection="row">
-                  <box width={16} height={1} backgroundColor={C.bgHighlight} paddingX={1}>
-                    <text>
-                      <span fg={C.green} bold>
-                        g
-                      </span>
-                      <span fg={C.white}> global</span>
-                    </text>
-                  </box>
-                  <box width={2} />
-                  <box width={16} height={1} paddingX={1}>
-                    <text>
-                      <span fg={C.cyan} bold>
-                        p
-                      </span>
-                      <span fg={C.fgMuted}> project (.claudish.json)</span>
-                    </text>
-                  </box>
-                </box>
-                <text>
-                  <span fg={C.green} bold>
-                    g{" "}
-                  </span>
-                  <span fg={C.fgMuted}>global · </span>
-                  <span fg={C.cyan} bold>
-                    p{" "}
-                  </span>
-                  <span fg={C.fgMuted}>project · </span>
-                  <span fg={C.red} bold>
-                    Esc{" "}
-                  </span>
-                  <span fg={C.fgMuted}>cancel</span>
-                </text>
-              </box>
-            )}
-
-            {/* Provider prefix picker */}
-            {mode === "pick_provider_prefix" && (
-              <box flexDirection="column">
-                {PROVIDER_PREFIXES.slice(0, 8).map((p, idx) => (
-                  <box
-                    key={p.name}
-                    height={1}
-                    backgroundColor={idx === providerPickerIndex ? C.bgHighlight : C.bg}
-                  >
-                    <text>
-                      <span fg={idx === providerPickerIndex ? C.white : C.dim}> </span>
-                      <span
-                        fg={idx === providerPickerIndex ? C.cyan : C.fgMuted}
-                        bold={idx === providerPickerIndex}
-                      >
-                        {p.prefix.padEnd(14).substring(0, 14)}
-                      </span>
-                      <span fg={C.dim}>{"  "}</span>
-                      <span fg={idx === providerPickerIndex ? C.fgMuted : C.dim}>
-                        {p.displayName}
-                      </span>
-                    </text>
-                  </box>
-                ))}
-                <text>
-                  <span fg={C.blue} bold>
-                    ↑↓{" "}
-                  </span>
-                  <span fg={C.fgMuted}>navigate · </span>
-                  <span fg={C.green} bold>
-                    Enter{" "}
-                  </span>
-                  <span fg={C.fgMuted}>select prefix · </span>
-                  <span fg={C.red} bold>
-                    Esc{" "}
-                  </span>
-                  <span fg={C.fgMuted}>back</span>
-                </text>
-              </box>
-            )}
-
-            {/* Normal text input (not scope/provider picker) */}
-            {mode !== "pick_profile_scope" && mode !== "pick_provider_prefix" && (
-              <box flexDirection="column">
-                <text>
-                  <span fg={C.green} bold>
-                    {"> "}
-                  </span>
-                  <span fg={editProfileValue === "auto" ? C.yellow : C.white}>
-                    {editProfileValue}
-                  </span>
-                  <span fg={C.cyan}>{"█"}</span>
-                </text>
-
-                {/* Suggestion list */}
-                {suggestions.length > 0 && (
-                  <box flexDirection="column">
-                    {suggestions.map((s, idx) => {
-                      const selected = idx === suggestionIndex;
-                      // Highlight matching portion
-                      const lower = editProfileValue.toLowerCase();
-                      const matchIdx = lower ? s.toLowerCase().indexOf(lower) : -1;
-                      return (
-                        <box key={s} height={1} backgroundColor={selected ? C.bgHighlight : C.bg}>
-                          <text>
-                            <span fg={selected ? C.dim : C.dim}>{"  "}</span>
-                            {matchIdx >= 0 && lower ? (
-                              <>
-                                <span fg={selected ? C.fgMuted : C.dim}>
-                                  {s.substring(0, matchIdx)}
-                                </span>
-                                <span fg={selected ? C.white : C.cyan} bold>
-                                  {s.substring(matchIdx, matchIdx + lower.length)}
-                                </span>
-                                <span fg={selected ? C.fgMuted : C.dim}>
-                                  {s.substring(matchIdx + lower.length)}
-                                </span>
-                              </>
-                            ) : (
-                              <span fg={selected ? C.white : C.fgMuted}>{s}</span>
-                            )}
-                          </text>
-                        </box>
-                      );
-                    })}
-                  </box>
-                )}
-
-                {editProfileValue === "auto" ? (
-                  <text>
-                    <span fg={C.yellow} bold>
-                      auto-route{" "}
-                    </span>
-                    <span fg={C.fgMuted}>— claudish will use the routing table · </span>
-                    <span fg={C.green} bold>
-                      Enter{" "}
-                    </span>
-                    <span fg={C.fgMuted}>to confirm · </span>
-                    <span fg={C.red} bold>
-                      Esc{" "}
-                    </span>
-                    <span fg={C.fgMuted}>cancel</span>
-                  </text>
-                ) : (
-                  <text>
-                    <span fg={C.green} bold>
-                      Enter{" "}
-                    </span>
-                    <span fg={C.fgMuted}>save · </span>
-                    <span fg={C.blue} bold>
-                      Tab{" "}
-                    </span>
-                    <span fg={C.fgMuted}>
-                      {editProfileValue === "" ? "pick provider · " : "autocomplete · "}
-                    </span>
-                    <span fg={C.blue} bold>
-                      ↑↓{" "}
-                    </span>
-                    <span fg={C.fgMuted}>suggestion · </span>
-                    <span fg={C.yellow} bold>
-                      a{" "}
-                    </span>
-                    <span fg={C.fgMuted}>auto-route · </span>
-                    <span fg={C.red} bold>
-                      Esc{" "}
-                    </span>
-                    <span fg={C.fgMuted}>cancel</span>
-                  </text>
-                )}
-              </box>
-            )}
-          </box>
-        )}
-      </box>
-    );
-  }
-
-  function ProfileDetail() {
-    const globalCfg = config;
-    const localCfg = loadLocalConfig();
-    const localProfileNames = localCfg
-      ? new Set(Object.keys(localCfg.profiles))
-      : new Set<string>();
-
-    // Resolve selected profile entry
-    const allEntries: Array<{
-      name: string;
-      scope: "local" | "global";
-      models: Record<string, string | undefined>;
-    }> = [];
-    if (localCfg) {
-      for (const [name, prof] of Object.entries(localCfg.profiles)) {
-        allEntries.push({ name, scope: "local", models: prof.models });
-      }
-    }
-    for (const [name, prof] of Object.entries(globalCfg.profiles)) {
-      allEntries.push({ name, scope: "global", models: prof.models });
-    }
-
-    const entry = allEntries[profileIndex];
-    const isActive = entry ? entry.name === globalCfg.defaultProfile : false;
-    const shadowed = entry ? entry.scope === "global" && localProfileNames.has(entry.name) : false;
-
-    return (
-      <box
-        height={DETAIL_H}
-        border
-        borderStyle="single"
-        borderColor={C.dim}
-        title={entry ? ` ${entry.name} ` : " (no selection) "}
-        backgroundColor={C.bgAlt}
-        flexDirection="column"
-        paddingX={1}
-      >
-        {entry ? (
-          <>
-            {(["opus", "sonnet", "haiku", "subagent"] as const).map((role) => {
-              const val = entry.models[role];
-              const isAuto = !val;
-              const label = role.padEnd(8);
-              return (
-                <text key={role}>
-                  <span fg={C.blue} bold>
-                    {label + ": "}
-                  </span>
-                  {isAuto ? (
-                    <>
-                      <span fg={C.yellow}>(auto-route</span>
-                      <span fg={C.dim}> — uses routing table</span>
-                      <span fg={C.yellow}>)</span>
-                    </>
-                  ) : (
-                    <span fg={C.cyan}>{val}</span>
-                  )}
-                </text>
-              );
-            })}
-            <text>
-              <span fg={C.blue} bold>
-                {"Scope:    "}
-              </span>
-              <span fg={entry.scope === "local" ? C.cyan : C.fgMuted}>
-                {entry.scope === "local"
-                  ? `local (.claudish.json)`
-                  : `global (~/.claudish/config.json)`}
-              </span>
-              {isActive && (
-                <span fg={C.orange} bold>
-                  {"  ● active"}
-                </span>
-              )}
-              {shadowed && <span fg={C.dim}>{"  (shadowed)"}</span>}
-            </text>
-          </>
-        ) : (
-          <text>
-            <span fg={C.fgMuted}>{"No profiles configured."}</span>
-          </text>
-        )}
-      </box>
-    );
-  }
-
-  // ── Routing tab ───────────────────────────────────────────────────────────
-
-  // Format a chain as inline text: "kimi → openrouter"
-  function chainStr(chain: string[]): string {
-    return chain.join(" → ");
-  }
-
-  // Reasons shown beneath each probe entry
-  const PROVIDER_REASONS: Record<string, string> = {
-    litellm: "LiteLLM proxy",
-    "opencode-zen": "Free tier (OpenCode Zen)",
-    "opencode-zen-go": "Zen Go plan",
-    kimi: "Native Kimi API",
-    "kimi-coding": "Kimi Coding Plan",
-    minimax: "Native MiniMax API",
-    "minimax-coding": "MiniMax Coding Plan",
-    glm: "Native GLM API",
-    "glm-coding": "GLM Coding Plan",
-    google: "Direct Gemini API",
-    openai: "Direct OpenAI API",
-    "openai-codex": "OpenAI Codex (Responses API)",
-    zai: "Z.AI API",
-    ollamacloud: "Cloud Ollama",
-    vertex: "Vertex AI Express",
-    openrouter: "Fallback: 580+ models",
-  };
-
-  function RoutingContent() {
-    // Full-screen probe takes over when not idle
-    const probeBoxH = contentH + DETAIL_H + 1; // spans content + detail area
-
-    if (probeMode === "input") {
-      return (
-        <box
-          height={probeBoxH}
-          border
-          borderStyle="single"
-          borderColor={C.focusBorder}
-          backgroundColor={C.bg}
-          flexDirection="column"
-          paddingX={2}
-          paddingY={1}
-        >
-          <text>
-            <span fg={C.white} bold>
-              {"Route Probe"}
-            </span>
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.fgMuted}>{"Enter a model name to trace its routing chain:"}</span>
-          </text>
-          <box flexDirection="row" height={1}>
-            <text>
-              <span fg={C.green} bold>
-                {"> "}
-              </span>
-              <span fg={C.white}>{probeModel}</span>
-              <span fg={C.cyan}>{"█"}</span>
-            </text>
-          </box>
-          <text> </text>
-          <text>
-            <span fg={C.dim}>{"Examples: kimi-k2  deepseek-r1  gemini-2.0-flash  gpt-4o"}</span>
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.fgMuted}>
-              {"The probe resolves the fallback chain and tests each provider's"}
-            </span>
-          </text>
-          <text>
-            <span fg={C.fgMuted}>{"API key in order, stopping at the first success."}</span>
-          </text>
-        </box>
-      );
-    }
-
-    if (probeMode === "running" || probeMode === "done") {
-      const successEntry = probeResults.find((e) => e.status === "success");
-      const allFailed = probeMode === "done" && !successEntry;
-      const totalMs = successEntry?.ms;
-
-      const statusBadge =
-        probeMode === "running"
-          ? { text: "probing...", color: C.yellow }
-          : successEntry
-            ? { text: "routed", color: C.green }
-            : { text: "no route", color: C.red };
-
-      return (
-        <box
-          height={probeBoxH}
-          border
-          borderStyle="single"
-          borderColor={probeMode === "running" ? C.focusBorder : C.blue}
-          backgroundColor={C.bg}
-          flexDirection="column"
-          paddingX={2}
-          paddingY={1}
-        >
-          {/* Title row */}
-          <box flexDirection="row" height={1}>
-            <text>
-              <span fg={C.white} bold>
-                {probeMode === "done" ? "Probe: " : "Probing: "}
-              </span>
-              <span fg={C.cyan} bold>
-                {probeModel}
-              </span>
-              <span fg={C.dim}>{"  "}</span>
-              {probeMode === "done" && (
-                <span fg={statusBadge.color} bold>
-                  {successEntry ? "● " : "✗ "}
-                  {statusBadge.text}
-                </span>
-              )}
-              {probeMode === "running" && <span fg={C.yellow}>{"◌ probing..."}</span>}
-            </text>
-          </box>
-          <text> </text>
-          {/* Route source */}
-          <text>
-            <span fg={C.fgMuted}>
-              {probeResults[0]?.reason ?? `Chain (${probeResults.length} providers):`}
-            </span>
-          </text>
-          <text> </text>
-          {/* Chain entries — 2 lines each */}
-          {probeResults.map((entry, idx) => {
-            const isNoKey = entry.status === "no_key";
-            const isNotReached = entry.status === "skipped";
-            const isSelected = entry.status === "success" && probeMode === "done";
-
-            const statusIcon =
-              entry.status === "success"
-                ? "●"
-                : entry.status === "failed"
-                  ? "✗"
-                  : entry.status === "testing"
-                    ? "◌"
-                    : isNoKey
-                      ? "○"
-                      : isNotReached
-                        ? "·"
-                        : "○";
-
-            const statusColor =
-              entry.status === "success"
-                ? C.green
-                : entry.status === "failed"
-                  ? C.red
-                  : entry.status === "testing"
-                    ? C.yellow
-                    : C.dim;
-
-            const nameCol = entry.displayName.padEnd(18).substring(0, 18);
-
-            const statusText =
-              entry.status === "success"
-                ? entry.ms !== undefined
-                  ? `${entry.ms}ms`
-                  : "success"
-                : entry.status === "failed"
-                  ? (entry.error ?? "failed")
-                  : entry.status === "testing"
-                    ? "testing..."
-                    : isNoKey
-                      ? "not configured, skipping"
-                      : isNotReached
-                        ? "not reached"
-                        : "waiting";
-
-            const reason = PROVIDER_REASONS[entry.provider] ?? entry.provider;
-
-            return (
-              <box key={entry.provider} flexDirection="column">
-                <text>
-                  <span fg={C.dim}>{`${idx + 1}. `}</span>
-                  <span
-                    fg={isNoKey ? C.dim : isSelected ? C.white : isNotReached ? C.dim : C.fgMuted}
-                    bold={isSelected}
-                  >
-                    {nameCol}
-                  </span>
-                  <span fg={C.dim}>{"  "}</span>
-                  <span fg={statusColor} bold={entry.status === "success"}>
-                    {statusIcon} {statusText}
-                  </span>
-                  {isSelected && (
-                    <span fg={C.green} bold>
-                      {" ← routed here"}
-                    </span>
-                  )}
-                </text>
-                <text>
-                  <span fg={C.dim}>{"    ↳ "}</span>
-                  <span fg={isNoKey ? C.dim : C.fgMuted}>{reason}</span>
-                </text>
-              </box>
-            );
-          })}
-          {/* Result line */}
-          {probeMode === "done" && (
-            <>
-              <text> </text>
-              <text>
-                {allFailed ? (
-                  <>
-                    <span fg={C.red} bold>
-                      {"Result: "}
-                    </span>
-                    <span fg={C.red}>{"✗ No provider could serve this model"}</span>
-                  </>
-                ) : (
-                  <>
-                    <span fg={C.green} bold>
-                      {"Result: "}
-                    </span>
-                    <span fg={C.fgMuted}>{"Routed to "}</span>
-                    <span fg={C.cyan} bold>
-                      {successEntry!.displayName}
-                    </span>
-                    {totalMs !== undefined && <span fg={C.fgMuted}>{` in ${totalMs}ms`}</span>}
-                  </>
-                )}
-              </text>
-            </>
-          )}
-        </box>
-      );
-    }
-
-    const innerH = contentH - 2;
-
-    return (
-      <box
-        height={contentH}
-        border
-        borderStyle="single"
-        borderColor={C.blue}
-        backgroundColor={C.bg}
-        flexDirection="column"
-        paddingX={1}
-      >
-        {/* Catch-all default — the only "global" default that actually exists.
-            Per-pattern defaults (gpt-* → codex/openai/openrouter, etc.) are
-            visible in the rule table below alongside any user overrides. */}
-        <text>
-          <span fg={C.blue} bold>
-            {" Catch-all default:"}
-          </span>
-          <span fg={C.fgMuted}>{"  (used for any model not matched by a rule)"}</span>
-        </text>
-        <text>
-          <span fg={C.dim}>{"  * "}</span>
-          <span fg={C.dim}>{"→ "}</span>
-          <span fg={C.cyan}>
-            {config.defaultProvider && config.defaultProvider.length > 0
-              ? config.defaultProvider
-              : (DEFAULT_ROUTING_RULES["*"]?.[0] ?? "openrouter")}
-          </span>
-          {config.defaultProvider && config.defaultProvider.length > 0 ? (
-            <span fg={C.fgMuted}>
-              {`  (defaultProvider — overrides built-in '${DEFAULT_ROUTING_RULES["*"]?.[0] ?? "openrouter"}')`}
-            </span>
-          ) : (
-            <span fg={C.fgMuted}>{"  (built-in)"}</span>
-          )}
-        </text>
-        <text>
-          <span fg={C.dim}>{" ─".repeat(Math.max(1, Math.floor((width - 6) / 2)))}</span>
-        </text>
-        {/* Custom rules header */}
-        <text>
-          <span fg={C.blue} bold>
-            {" Custom rules:"}
-          </span>
-          <span fg={C.fgMuted}>{"  (override default for matching models)"}</span>
-        </text>
-        {/* Custom rules or empty state */}
-        {ruleEntries.length === 0 && !isRoutingInput && (
-          <text>
-            <span fg={C.fgMuted}>{" None configured. Press "}</span>
-            <span fg={C.green} bold>
-              a
-            </span>
-            <span fg={C.fgMuted}>{" to add."}</span>
-          </text>
-        )}
-        {ruleEntries.length > 0 && (
-          <>
-            <text>
-              <span fg={C.blue} bold>
-                {"PATTERN         "}
-              </span>
-              <span fg={C.blue} bold>
-                {"CHAIN"}
-              </span>
-            </text>
-            {ruleEntries.slice(0, Math.max(0, innerH - 3)).map(([pat, chain], idx) => {
-              const sel = idx === providerIndex;
-              return (
-                <box
-                  key={pat}
-                  height={1}
-                  flexDirection="row"
-                  backgroundColor={sel ? C.bgHighlight : C.bg}
-                >
-                  <text>
-                    <span fg={sel ? C.white : C.fgMuted} bold={sel}>
-                      {pat.padEnd(16).substring(0, 16)}
-                    </span>
-                    <span fg={C.dim}>{"  "}</span>
-                    <span fg={sel ? C.cyan : C.fgMuted}>{chainStr(chain)}</span>
-                  </text>
-                </box>
-              );
-            })}
-          </>
-        )}
-
-        {/* Input fields */}
-        {mode === "add_routing_pattern" && (
-          <box flexDirection="column">
-            <text>
-              <span fg={C.blue} bold>
-                {"Pattern "}
-              </span>
-              <span fg={C.dim}>{"(e.g. kimi-*, gpt-4o):"}</span>
-            </text>
-            <text>
-              <span fg={C.green} bold>
-                {"> "}
-              </span>
-              <span fg={C.white}>{routingPattern}</span>
-              <span fg={C.cyan}>{"█"}</span>
-            </text>
-            <text>
-              <span fg={C.green} bold>
-                Enter{" "}
-              </span>
-              <span fg={C.fgMuted}>to continue · </span>
-              <span fg={C.red} bold>
-                Esc{" "}
-              </span>
-              <span fg={C.fgMuted}>to cancel</span>
-            </text>
-          </box>
-        )}
-        {mode === "add_routing_chain" && (
-          <box flexDirection="column">
-            <text>
-              <span fg={C.blue} bold>
-                {"Select providers for "}
-              </span>
-              <span fg={C.white} bold>
-                {routingPattern}
-              </span>
-              <span fg={C.dim}>{" (Space=toggle, 1-9=set position, Enter=save)"}</span>
-            </text>
-            {chainOrder.length > 0 && (
-              <text>
-                <span fg={C.fgMuted}>{"  Chain: "}</span>
-                <span fg={C.cyan}>{chainOrder.join(" → ")}</span>
-              </text>
-            )}
-            {CHAIN_PROVIDERS.map((prov, idx) => {
-              const isCursor = idx === chainCursor;
-              const isOn = chainSelected.has(prov.name);
-              const pos = isOn ? chainOrder.indexOf(prov.name) + 1 : 0;
-              const hasKey = !!(
-                config.apiKeys?.[prov.apiKeyEnvVar] || process.env[prov.apiKeyEnvVar]
-              );
-              const label = prov.displayName.padEnd(18).substring(0, 18);
-              return (
-                <box key={prov.name} height={1} backgroundColor={isCursor ? C.bgHighlight : C.bg}>
-                  <text>
-                    {isOn ? (
-                      <span fg={C.green} bold>{` [${pos}] `}</span>
-                    ) : (
-                      <span fg={C.dim}>{" [ ] "}</span>
-                    )}
-                    <span fg={isCursor ? C.white : hasKey ? C.fgMuted : C.dim} bold={isCursor}>
-                      {label}
-                    </span>
-                    {hasKey ? (
-                      <span fg={C.green}>{" ●"}</span>
-                    ) : (
-                      <span fg={C.dim}>{" ○ no key"}</span>
-                    )}
-                  </text>
-                </box>
-              );
-            })}
-          </box>
-        )}
-      </box>
-    );
-  }
-
-  function RoutingDetail() {
-    // Probe is full-screen — no separate detail panel shown
-    if (probeMode !== "idle") {
-      return null;
-    }
-
-    return (
-      <box
-        height={DETAIL_H}
-        border
-        borderStyle="single"
-        borderColor={C.dim}
-        title=" Examples "
-        backgroundColor={C.bgAlt}
-        flexDirection="column"
-        paddingX={1}
-      >
-        <text>
-          <span fg={C.fgMuted}>{"  kimi-*      "}</span>
-          <span fg={C.dim}>{" → "}</span>
-          <span fg={C.cyan}>{"kimi, openrouter"}</span>
-        </text>
-        <text>
-          <span fg={C.fgMuted}>{"  gpt-*       "}</span>
-          <span fg={C.dim}>{" → "}</span>
-          <span fg={C.cyan}>{"oai, litellm"}</span>
-        </text>
-        <text>
-          <span fg={C.fgMuted}>{"  gemini-*    "}</span>
-          <span fg={C.dim}>{" → "}</span>
-          <span fg={C.cyan}>{"google, zen, openrouter"}</span>
-        </text>
-        <text>
-          <span fg={C.fgMuted}>{"  deepseek-*  "}</span>
-          <span fg={C.dim}>{" → "}</span>
-          <span fg={C.cyan}>{"zen, openrouter"}</span>
-        </text>
-        <text>
-          <span fg={C.dim}>{"  Glob pattern (* = any). Chain tried left to right. "}</span>
-          <span fg={C.cyan} bold>
-            {ruleEntries.length}
-          </span>
-          <span fg={C.fgMuted}>
-            {" custom rule"}
-            {ruleEntries.length !== 1 ? "s" : ""}
-          </span>
-        </text>
-      </box>
-    );
-  }
-
-  // ── Privacy tab ───────────────────────────────────────────────────────────
-  function PrivacyContent() {
-    const halfW = Math.floor((width - 4) / 2);
-    const cardH = Math.max(7, contentH - 1);
-
-    return (
-      <box height={contentH} flexDirection="row" backgroundColor={C.bg} paddingX={1}>
-        {/* Telemetry card */}
-        <box
-          width={halfW}
-          height={cardH}
-          border
-          borderStyle="single"
-          borderColor={activeTab === "privacy" ? C.blue : C.dim}
-          title=" Telemetry "
-          backgroundColor={C.bg}
-          flexDirection="column"
-          paddingX={1}
-        >
-          <text>
-            <span fg={C.blue} bold>
-              Status:{" "}
-            </span>
-            {telemetryEnabled ? (
-              <span fg={C.green} bold>
-                ● Enabled
-              </span>
-            ) : (
-              <span fg={C.fgMuted}>○ Disabled</span>
-            )}
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.fgMuted}>Collects anonymized platform info and</span>
-          </text>
-          <text>
-            <span fg={C.fgMuted}>sanitized error types to improve claudish.</span>
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.white} bold>
-              Never sends keys, prompts, or paths.
-            </span>
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.dim}>Press [</span>
-            <span fg={C.green} bold>
-              t
-            </span>
-            <span fg={C.dim}>] to toggle.</span>
-          </text>
-        </box>
-
-        {/* Usage stats card */}
-        <box
-          width={width - 4 - halfW}
-          height={cardH}
-          border
-          borderStyle="single"
-          borderColor={activeTab === "privacy" ? C.blue : C.dim}
-          title=" Usage Stats "
-          backgroundColor={C.bg}
-          flexDirection="column"
-          paddingX={1}
-        >
-          <text>
-            <span fg={C.blue} bold>
-              Status:{" "}
-            </span>
-            {statsEnabled ? (
-              <span fg={C.green} bold>
-                ● Enabled
-              </span>
-            ) : (
-              <span fg={C.fgMuted}>○ Disabled</span>
-            )}
-          </text>
-          <text>
-            <span fg={C.blue} bold>
-              Buffer:{" "}
-            </span>
-            <span fg={C.white} bold>
-              {bufStats.events}
-            </span>
-            <span fg={C.fgMuted}> events (</span>
-            <span fg={C.yellow}>{bytesHuman(bufStats.bytes)}</span>
-            <span fg={C.fgMuted}>)</span>
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.fgMuted}>Collects local, anonymous stats on model</span>
-          </text>
-          <text>
-            <span fg={C.fgMuted}>usage, latency, and token counts.</span>
-          </text>
-          <text> </text>
-          <text>
-            <span fg={C.dim}>Press [</span>
-            <span fg={C.green} bold>
-              u
-            </span>
-            <span fg={C.dim}>] to toggle, [</span>
-            <span fg={C.red} bold>
-              c
-            </span>
-            <span fg={C.dim}>] to clear buffer.</span>
-          </text>
-        </box>
-      </box>
-    );
-  }
-
-  function PrivacyDetail() {
-    return (
-      <box
-        height={DETAIL_H}
-        border
-        borderStyle="single"
-        borderColor={C.dim}
-        title=" Your Privacy "
-        backgroundColor={C.bgAlt}
-        flexDirection="column"
-        paddingX={1}
-      >
-        <text>
-          <span fg={C.fgMuted}>
-            Telemetry and usage stats are always opt-in and never send personally identifiable data.
-          </span>
-        </text>
-        <text>
-          <span fg={C.fgMuted}>
-            All data is anonymized before transmission. You can disable either independently.
-          </span>
-        </text>
-      </box>
-    );
-  }
-
-  // ── Footer hotkeys ────────────────────────────────────────────────────────
-  function Footer() {
-    let keys: Array<[string, string, string]>;
-    if (activeTab === "routing" && probeMode === "input") {
-      keys = [
-        [C.green, "Enter", "probe"],
-        [C.red, "Esc", "cancel"],
-      ];
-    } else if (activeTab === "routing" && probeMode === "running") {
-      keys = [
-        [C.yellow, "◌", "probing..."],
-        [C.red, "Esc", "cancel"],
-      ];
-    } else if (activeTab === "routing" && probeMode === "done") {
-      keys = [
-        [C.cyan, "p", "back to routes"],
-        [C.green, "Enter", "probe another"],
-        [C.red, "Esc", "back to routes"],
-        [C.dim, "q", "quit"],
-      ];
-    } else if (activeTab === "providers") {
-      keys = [
-        [C.blue, "↑↓", "navigate"],
-        [C.green, "s", "set key"],
-        [C.green, "e", "endpoint"],
-        [C.cyan, "t", "test key"],
-        [C.red, "x", "remove"],
-        [C.blue, "Tab", "section"],
-        [C.dim, "q", "quit"],
-      ];
-    } else if (activeTab === "profiles" && mode === "pick_profile_scope") {
-      keys = [
-        [C.green, "g", "global"],
-        [C.cyan, "p", "project"],
-        [C.red, "Esc", "cancel"],
-      ];
-    } else if (activeTab === "profiles" && mode === "pick_provider_prefix") {
-      keys = [
-        [C.blue, "↑↓", "navigate"],
-        [C.green, "Enter", "select prefix"],
-        [C.red, "Esc", "back"],
-      ];
-    } else if (activeTab === "profiles" && isProfileEditMode) {
-      keys = [
-        [C.green, "Enter", "save field"],
-        [C.blue, "Tab", "provider picker"],
-        [C.blue, "↑↓", "suggestion"],
-        [C.yellow, "a", "auto-route"],
-        [C.red, "Esc", "cancel"],
-      ];
-    } else if (activeTab === "profiles") {
-      keys = [
-        [C.blue, "↑↓", "navigate"],
-        [C.green, "Enter", "activate"],
-        [C.cyan, "n", "new"],
-        [C.green, "e", "edit"],
-        [C.red, "d", "delete"],
-        [C.blue, "Tab", "section"],
-        [C.dim, "q", "quit"],
-      ];
-    } else if (activeTab === "routing") {
-      keys = [
-        [C.blue, "↑↓", "navigate"],
-        [C.green, "a", "add rule"],
-        [C.red, "d", "delete"],
-        [C.cyan, "p", "probe"],
-        [C.blue, "Tab", "section"],
-        [C.dim, "q", "quit"],
-      ];
-    } else {
-      keys = [
-        [C.green, "t", "telemetry"],
-        [C.green, "u", "stats"],
-        [C.red, "c", "clear"],
-        [C.blue, "Tab", "section"],
-        [C.dim, "q", "quit"],
-      ];
-    }
-
-    return (
-      <box height={FOOTER_H} flexDirection="row" paddingX={1} backgroundColor={C.bgAlt}>
-        <text>
-          {keys.map(([color, key, label], i) => (
-            <span key={i}>
-              {i > 0 && <span fg={C.dim}>{" │ "}</span>}
-              <span fg={color as string} bold>
-                {key}
-              </span>
-              <span fg={C.fgMuted}> {label}</span>
-            </span>
-          ))}
-        </text>
-      </box>
-    );
-  }
 
   // ── Main render ───────────────────────────────────────────────────────────
   return (
@@ -2313,36 +863,97 @@ export function App() {
       </box>
 
       {/* Tab bar */}
-      <TabBar />
+      <TabBar activeTab={activeTab} statusMsg={statusMsg} width={width} />
 
       {/* Content + detail */}
       {activeTab === "providers" && (
         <>
-          <ProvidersContent />
-          <ProviderDetail />
+          <ProvidersContent
+            config={config}
+            displayProviders={displayProviders}
+            providerIndex={providerIndex}
+            testResults={testResults}
+            width={width}
+            contentH={contentH}
+            isInputMode={isInputMode}
+          />
+          <ProviderDetail
+            selectedProvider={selectedProvider}
+            mode={mode}
+            inputValue={inputValue}
+            setInputValue={setInputValue}
+            width={width}
+            hasCfgKey={hasCfgKey}
+            hasEnvKey={hasEnvKey}
+            hasKey={hasKey}
+            cfgKeyMask={cfgKeyMask}
+            envKeyMask={envKeyMask}
+            keySrc={keySrc}
+            activeEndpoint={activeEndpoint}
+            testResults={testResults}
+            isInputMode={isInputMode}
+          />
         </>
       )}
       {activeTab === "profiles" && (
         <>
-          <ProfilesContent />
-          <ProfileDetail />
+          <ProfilesContent
+            config={config}
+            activeTab={activeTab}
+            mode={mode}
+            profileScope={profileScope}
+            profileIndex={profileIndex}
+            editProfileName={editProfileName}
+            editProfileValue={editProfileValue}
+            suggestions={suggestions}
+            suggestionIndex={suggestionIndex}
+            providerPickerIndex={providerPickerIndex}
+            width={width}
+            contentH={contentH}
+          />
+          <ProfileDetail config={config} profileIndex={profileIndex} />
         </>
       )}
       {activeTab === "routing" && (
         <>
-          <RoutingContent />
-          <RoutingDetail />
+          <RoutingContent
+            config={config}
+            probeMode={probeMode}
+            probeModel={probeModel}
+            probeResults={probeResults}
+            mode={mode}
+            routingPattern={routingPattern}
+            chainSelected={chainSelected}
+            chainOrder={chainOrder}
+            chainCursor={chainCursor}
+            // NOTE: `providerIndex` is shared with the Providers tab here. See
+            // "Known wart" in ai-docs/app-tsx-split/walkthrough.md — switching
+            // tabs preserves the cursor across two unrelated lists.
+            providerIndex={providerIndex}
+            ruleEntries={ruleEntries}
+            width={width}
+            contentH={contentH}
+            isRoutingInput={isRoutingInput}
+          />
+          <RoutingDetail probeMode={probeMode} ruleEntries={ruleEntries} />
         </>
       )}
       {activeTab === "privacy" && (
         <>
-          <PrivacyContent />
+          <PrivacyContent
+            activeTab={activeTab}
+            telemetryEnabled={telemetryEnabled}
+            statsEnabled={statsEnabled}
+            bufStats={bufStats}
+            width={width}
+            contentH={contentH}
+          />
           <PrivacyDetail />
         </>
       )}
 
       {/* Footer */}
-      <Footer />
+      <Footer activeTab={activeTab} mode={mode} probeMode={probeMode} />
     </box>
   );
 }
