+++ packages/cli/src/tui/components/ProfileDetail.tsx
@@ -1,0 +1,96 @@
+/** @jsxImportSource @opentui/react */
+import { C } from "../theme.js";
+import { DETAIL_H } from "../constants.js";
+import { loadLocalConfig } from "../../profile-config.js";
+import type { ClaudishProfileConfig } from "../../profile-config.js";
+
+interface ProfileDetailProps {
+  config: ClaudishProfileConfig;
+  profileIndex: number;
+}
+
+export function ProfileDetail({ config, profileIndex }: ProfileDetailProps) {
+  const globalCfg = config;
+  const localCfg = loadLocalConfig();
+  const localProfileNames = localCfg
+    ? new Set(Object.keys(localCfg.profiles))
+    : new Set<string>();
+
+  // Resolve selected profile entry
+  const allEntries: Array<{
+    name: string;
+    scope: "local" | "global";
+    models: Record<string, string | undefined>;
+  }> = [];
+  if (localCfg) {
+    for (const [name, prof] of Object.entries(localCfg.profiles)) {
+      allEntries.push({ name, scope: "local", models: prof.models });
+    }
+  }
+  for (const [name, prof] of Object.entries(globalCfg.profiles)) {
+    allEntries.push({ name, scope: "global", models: prof.models });
+  }
+
+  const entry = allEntries[profileIndex];
+  const isActive = entry ? entry.name === globalCfg.defaultProfile : false;
+  const shadowed = entry ? entry.scope === "global" && localProfileNames.has(entry.name) : false;
+
+  return (
+    <box
+      height={DETAIL_H}
+      border
+      borderStyle="single"
+      borderColor={C.dim}
+      title={entry ? ` ${entry.name} ` : " (no selection) "}
+      backgroundColor={C.bgAlt}
+      flexDirection="column"
+      paddingX={1}
+    >
+      {entry ? (
+        <>
+          {(["opus", "sonnet", "haiku", "subagent"] as const).map((role) => {
+            const val = entry.models[role];
+            const isAuto = !val;
+            const label = role.padEnd(8);
+            return (
+              <text key={role}>
+                <span fg={C.blue} bold>
+                  {label + ": "}
+                </span>
+                {isAuto ? (
+                  <>
+                    <span fg={C.yellow}>(auto-route</span>
+                    <span fg={C.dim}> — uses routing table</span>
+                    <span fg={C.yellow}>)</span>
+                  </>
+                ) : (
+                  <span fg={C.cyan}>{val}</span>
+                )}
+              </text>
+            );
+          })}
+          <text>
+            <span fg={C.blue} bold>
+              {"Scope:    "}
+            </span>
+            <span fg={entry.scope === "local" ? C.cyan : C.fgMuted}>
+              {entry.scope === "local"
+                ? `local (.claudish.json)`
+                : `global (~/.claudish/config.json)`}
+            </span>
+            {isActive && (
+              <span fg={C.orange} bold>
+                {"  ● active"}
+              </span>
+            )}
+            {shadowed && <span fg={C.dim}>{"  (shadowed)"}</span>}
+          </text>
+        </>
+      ) : (
+        <text>
+          <span fg={C.fgMuted}>{"No profiles configured."}</span>
+        </text>
+      )}
+    </box>
+  );
+}
