--- packages/cli/src/tui/panels/StatsPanel.tsx
@@ -1,3 +1,0 @@
-/** @jsxImportSource @opentui/react */
-// Replaced by App.tsx dashboard
-export {};
