--- packages/cli/src/tui/panels/TelemetryPanel.tsx
@@ -1,3 +1,0 @@
-/** @jsxImportSource @opentui/react */
-// Replaced by App.tsx dashboard
-export {};
